Clean up tests. Restructure DI and app config. Removed angularfire library, replaced with Rxjs / standard firestore.
Fixed load of test data using mock data stored in project. Load this data to the emulator for affected tests. Other tests use stub services to return data without contacting the emulator or real databases. Cleaned up FirebaseAppService test and tested for correctness and failing when the local emulator is down or has not been started.